test(Types): consolidate RequiredKeys coverage
